+++ regression/reference/generic/wrapgeneric.h
@@ -48,8 +48,6 @@ void GEN_SavePointer_float2d_bufferify(float * addr, int type,
     size_t size);
 #endif
 
-void GEN_SavePointer2(void * addr, int type, size_t size);
-
 void GEN_SavePointer2_float1d_bufferify(float * addr, int type,
     size_t size);
 
@@ -66,9 +64,8 @@ void GEN_GetPointerAsPointer_float2d_bufferify(
     GEN_SHROUD_array *SHT_addr_cdesc);
 #endif
 
-GEN_StructAsClass * GEN_CreateStructAsClass(GEN_StructAsClass * SHC_rv);
-
-long GEN_UpdateStructAsClass(GEN_StructAsClass * arg, long inew);
+GEN_StructAsClass * GEN_CreateStructAsClass_bufferify(
+    GEN_StructAsClass * SHC_rv);
 
 long GEN_UpdateStructAsClass_int_bufferify(GEN_StructAsClass * arg,
     int inew);