+++ regression/reference/struct-c/wrapstruct.h
@@ -26,44 +26,38 @@
 int STR_passStruct2_bufferify(const Cstruct1 * s1, char *outbuf,
     int SHT_outbuf_len);
 
-void STR_returnStructByValue(int i, double d, Cstruct1 *SHC_rv);
-
-Cstruct1 * STR_returnStructPtr1(int i, double d);
+void STR_returnStructByValue_bufferify(int i, double d,
+    Cstruct1 *SHC_rv);
 
 void STR_returnStructPtr1_bufferify(int i, double d,
     STR_SHROUD_array *SHT_rv_cdesc);
 
-Cstruct1 * STR_returnStructPtr2(int i, double d, char * outbuf);
-
 void STR_returnStructPtr2_bufferify(int i, double d, char *outbuf,
     int SHT_outbuf_len, STR_SHROUD_array *SHT_rv_cdesc);
 
-Cstruct1 * STR_returnStructPtrArray(void);
-
 void STR_returnStructPtrArray_bufferify(STR_SHROUD_array *SHT_rv_cdesc);
-
-Cstruct_list * STR_get_global_struct_list(void);
 
 void STR_get_global_struct_list_bufferify(
     STR_SHROUD_array *SHT_rv_cdesc);
 
-STR_Cstruct_as_class * STR_Create_Cstruct_as_class(
+STR_Cstruct_as_class * STR_Create_Cstruct_as_class_bufferify(
     STR_Cstruct_as_class * SHC_rv);
 
-STR_Cstruct_as_class * STR_Create_Cstruct_as_class_args(int x, int y,
+STR_Cstruct_as_class * STR_Create_Cstruct_as_class_args_bufferify(int x,
+    int y, STR_Cstruct_as_class * SHC_rv);
+
+void STR_Return_Cstruct_as_class_bufferify(
     STR_Cstruct_as_class * SHC_rv);
 
-void STR_Return_Cstruct_as_class(STR_Cstruct_as_class * SHC_rv);
-
-void STR_Return_Cstruct_as_class_args(int x, int y,
+void STR_Return_Cstruct_as_class_args_bufferify(int x, int y,
     STR_Cstruct_as_class * SHC_rv);
 
-int STR_Cstruct_as_class_sum(STR_Cstruct_as_class * point);
+int STR_Cstruct_as_class_sum_bufferify(STR_Cstruct_as_class * point);
 
-STR_Cstruct_as_subclass * STR_Create_Cstruct_as_subclass_args(int x,
-    int y, int z, STR_Cstruct_as_subclass * SHC_rv);
+STR_Cstruct_as_subclass * STR_Create_Cstruct_as_subclass_args_bufferify(
+    int x, int y, int z, STR_Cstruct_as_subclass * SHC_rv);
 
-void STR_Return_Cstruct_as_subclass_args(int x, int y, int z,
+void STR_Return_Cstruct_as_subclass_args_bufferify(int x, int y, int z,
     STR_Cstruct_as_subclass * SHC_rv);
 
 const double * STR_Cstruct_ptr_get_const_dvalue(Cstruct_ptr * SH_this);